Cleans up backend API doc
The backend API documentation sounded like an internal document so
far. This commit adds some links to the other documentation and
points out that the backend API is used by the CLI tool which is
a more convenient tool for managing widgets on Contentful. So
developers can stop reading there and follow one of the links to
more useful/less confusing information.